+++ src/main/java/ddd/caffeine/ratrip/module/place/presentation/PlaceController.java
@@ -11,9 +11,11 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -26,7 +28,9 @@ import ddd.caffeine.ratrip.module.place.presentation.dto.detail.PlaceDetailsResp
 import ddd.caffeine.ratrip.module.place.presentation.dto.search.PlaceSearchRequestDto;
 import ddd.caffeine.ratrip.module.place.presentation.dto.search.PlaceSearchResponseDto;
 import ddd.caffeine.ratrip.module.user.domain.User;
+import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
 
 /**
@@ -79,4 +83,24 @@ public class PlaceController {
 		PlaceInRegionResponseDto response = placeService.readPlacesInRegionsApi(regions, pageable);
 		return ResponseEntity.ok(response);
 	}
+
+	@Operation(summary = "[인증] 북마크 추가")
+	@ApiResponse(description = "북마크 추가 성공 시, 북마크 ID 반환")
+	@PostMapping("/{id}/bookmark")
+	public ResponseEntity<java.util.UUID> callAddBookmarkApi(
+		@PathVariable @UUID @NotEmpty String id,
+		@Parameter(hidden = true) @AuthenticationPrincipal User user) {
+		java.util.UUID response = placeService.addBookMark(java.util.UUID.fromString(id), user);
+
+		return ResponseEntity.ok(response);
+	}
+
+	@Operation(summary = "[인증] 북마크 삭제")
+	@DeleteMapping("/{id}")
+	public ResponseEntity<String> deleteBookmark(
+		@PathVariable @UUID @NotEmpty String id,
+		@Parameter(hidden = true) @AuthenticationPrincipal User user) {
+		return ResponseEntity.ok();
+	}
+
 }
